requirements.txt
requests
python-dotenv
ddgs
google-genai
tavily-python
crawlee[beautifulsoup]
